# soroban-rpc: getLedgerEntries: query multiple ledger entries at once