--- a/cmd/soroban-rpc/internal/db/ledgerentry_test.go
+++ b/cmd/soroban-rpc/internal/db/ledgerentry_test.go
@@ -736,6 +736,65 @@ forloop:
 
 }
 
+func benchmarkLedgerEntry(b *testing.B, cached bool, includeExpired bool) {
+	db := NewTestDB(b)
+	keyUint32 := xdr.Uint32(0)
+	data := xdr.ContractDataEntry{
+		Contract: xdr.ScAddress{
+			Type:       xdr.ScAddressTypeScAddressTypeContract,
+			ContractId: &xdr.Hash{0xca, 0xfe},
+		},
+		Key: xdr.ScVal{
+			Type: xdr.ScValTypeScvU32,
+			U32:  &keyUint32,
+		},
+		Durability: xdr.ContractDataDurabilityPersistent,
+		Body: xdr.ContractDataEntryBody{
+			BodyType: xdr.ContractEntryBodyTypeDataEntry,
+			Data: &xdr.ContractDataEntryData{
+				Val: xdr.ScVal{
+					Type: xdr.ScValTypeScvU32,
+					U32:  &keyUint32,
+				},
+			},
+		},
+		ExpirationLedgerSeq: math.MaxUint32,
+	}
+	key, entry := getContractDataLedgerEntry(b, data)
+	tx, err := NewReadWriter(db, 150, 15).NewTx(context.Background())
+	assert.NoError(b, err)
+	assert.NoError(b, tx.LedgerEntryWriter().UpsertLedgerEntry(entry))
+	assert.NoError(b, tx.Commit(2))
+	reader := NewLedgerEntryReader(db)
+	const numQueriesPerOp = 15
+	b.ResetTimer()
+	b.StopTimer()
+	for i := 0; i < b.N; i++ {
+		var readTx LedgerEntryReadTx
+		var err error
+		if cached {
+			readTx, err = reader.NewCachedTx(context.Background())
+		} else {
+			readTx, err = reader.NewTx(context.Background())
+		}
+		assert.NoError(b, err)
+		for i := 0; i < numQueriesPerOp; i++ {
+			b.StartTimer()
+			found, _, err := GetLedgerEntry(readTx, includeExpired, key)
+			b.StopTimer()
+			assert.NoError(b, err)
+			assert.True(b, found)
+		}
+		assert.NoError(b, readTx.Done())
+	}
+}
+
+func BenchmarkGetLedgerEntry(b *testing.B) {
+	b.Run("With cache, include expired", func(b *testing.B) { benchmarkLedgerEntry(b, true, true) })
+	b.Run("With cache, exclude expired", func(b *testing.B) { benchmarkLedgerEntry(b, true, false) })
+	b.Run("Without cache, exclude expired", func(b *testing.B) { benchmarkLedgerEntry(b, false, false) })
+}
+
 func BenchmarkLedgerUpdate(b *testing.B) {
 	db := NewTestDB(b)
 	keyUint32 := xdr.Uint32(0)
@@ -773,7 +832,6 @@ func BenchmarkLedgerUpdate(b *testing.B) {
 		}
 		assert.NoError(b, tx.Commit(uint32(i+1)))
 	}
-	b.StopTimer()
 }
 
 func NewTestDB(tb testing.TB) *DB {
@@ -783,9 +841,6 @@ func NewTestDB(tb testing.TB) *DB {
 	if err != nil {
 		assert.NoError(tb, db.Close())
 	}
-	var ver []string
-	assert.NoError(tb, db.SelectRaw(context.Background(), &ver, "SELECT sqlite_version()"))
-	tb.Logf("using sqlite version: %v", ver)
 	tb.Cleanup(func() {
 		assert.NoError(tb, db.Close())
 	})