This improves the performance of the `getLedgerEntries` method and also paves the way to querying multiple entries from libpreflight (ahead of the upcoming ledger expiration changes)