@2opremio 2023-08-24
I think I am happy with this.