@tsachiherman 2023-08-24
> BTW, we should probably limit the number of entries requested at once.
> 
> @tsachiherman did you get to add a limit to the size of JSON requests? We should also do that.

I'm not sure where that would be. But we should definitely add a limit to the number of keys that can be queried in a single getLedgerEntries call.

Adding a limit to the json size request is a bit harder to implement; we might be able to limit the http request size by adding a middleware.