@2opremio 2023-08-24
> BTW, we should probably limit the number of entries requested at once.
> 
> @tsachiherman did you get to add a limit to the size of JSON requests? We should also do that.

@tsachiherman did you note this comments?