@tsachiherman 2023-08-24
> > BTW, we should probably limit the number of entries requested at once.
> > @tsachiherman did you get to add a limit to the size of JSON requests? We should also do that.
> 
> @tsachiherman did you note this comments?

I agree that we should limit the number of entries requested.

As for the other one, I think that it's doable via a middleware - but it's more work .. with less obvious gain. ( unless it's easy to configure )