@2opremio 2023-08-24
Limiting the size of http requests a pretty common practice.

Should we limit the number of entries queried to, say, 200?